Disable SecurityConfig for tests, fix "value-paths" in properties
