+++ authentication/src/main/resources/application.properties
@@ -2,7 +2,7 @@ spring.datasource.url=jdbc:postgresql://main-service-db:5432/db_movies
 spring.datasource.username=main
 spring.datasource.password=password
 spring.datasource.driver-class-name=org.postgresql.Driver
-user.service.url=http://user-service/api/v1/users
-privilege.service.url=http://user-service/api/v1/privileges
+user.service.url=http://user-service:8080/api/v1/users
+privilege.service.url=http://user-service:8080/api/v1/privileges
 jwt.secret=very_long_text
 jwt.validity=43200000
