+++ movies-main/src/main/resources/application.properties
@@ -1,0 +1,2 @@
+user.service.url=http://user-service:8080/api/v1/users
+privilege.service.url=http://user-service:8080/api/v1/privileges
