+++ movies-main/src/main/resources/application.yml
@@ -16,4 +16,4 @@ server:
 
 springdoc:
   swagger-ui:
-    path: /swagger-ui
+    enabled: true
